+++ app/lib/ui/guidelines/guidelines.dart
@@ -12,18 +12,31 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/ods_flutter_app_localizations.dart';
-import 'package:ods_flutter_demo/ui/guidelines/colours/guideline_colours_screen.dart';
+import 'package:ods_flutter_demo/ui/guidelines/colours/palette/coulour_palette.dart';
+import 'package:ods_flutter_demo/ui/guidelines/colours/theme/colour_theme.dart';
 import 'package:ods_flutter_demo/ui/guidelines/guidelines_entities.dart';
 import 'package:ods_flutter_demo/ui/guidelines/spacings/guideline_spacings_screen.dart';
 import 'package:ods_flutter_demo/ui/guidelines/typography/guideline_typography_screen.dart';
 
 List<Guideline> guidelines(BuildContext context) {
   return [
-    Guideline(
-        AppLocalizations.of(context)!.guidelinesColours,
-        'assets/il_color.png',
-        AppLocalizations.of(context)!.cardSmallVariantTitle,
-        GuidelineColoursScreen()),
+    Guideline.withVariant(
+      AppLocalizations.of(context)!.guidelinesColours,
+      'assets/il_color.png',
+      AppLocalizations.of(context)!.guidelinesColoursDescription,
+      [
+        Variant(
+          AppLocalizations.of(context)!.guidelinesColoursThemeVariantTitle,
+          AppLocalizations.of(context)!.guidelinesColoursThemeVariantSubtitle,
+          ColourTheme(),
+        ),
+        Variant(
+          AppLocalizations.of(context)!.guidelinesColoursPaletteVariantTitle,
+          AppLocalizations.of(context)!.guidelinesColoursPaletteVariantSubtitle,
+          ColourPalette(),
+        ),
+      ],
+    ),
     Guideline(
       AppLocalizations.of(context)!.guidelinesSpacings,
       'assets/il_spacing.png',